requirements.txt
#pip freeze>requirements.txt
#pip install -r requirements.txt
azure-ai-documentintelligence==1.0.0b3
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
gunicorn==23.0.0
idna==3.8
importlib_metadata==8.4.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
msrest==0.7.1
numpy==2.0.2
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
zipp==3.20.1